+++ CHANGELOG.rst
@@ -9,6 +9,8 @@ Unreleased
 ----------
 Changed
 
+- Changed rows and column in outlier-table
+
 - Added TODO List
 
 - Updated functionality to retrieve the stacktrace of an Outlier